rename tokenpicker
