+++ apps/hyperdrive-trading/src/ui/token/TokenPicker.tsx
@@ -1,13 +1,18 @@
 import { TokenConfig } from "@delvtech/hyperdrive-appconfig";
-import { ChevronDownIcon } from "@heroicons/react/24/outline";
+import {
+  ChevronDownIcon,
+  MagnifyingGlassIcon,
+  XMarkIcon,
+} from "@heroicons/react/20/solid";
 import classNames from "classnames";
-import { ReactElement } from "react";
+import Fuse from "fuse.js";
+import { ReactElement, useMemo, useRef, useState } from "react";
 import { formatBalance } from "src/ui/base/formatting/formatBalance";
 import { Address } from "viem";
 
 export interface TokenChoice {
-  disabled?: boolean;
   tokenConfig: TokenConfig;
+  disabled?: boolean;
   tokenBalance?: bigint;
 }
 
@@ -16,13 +21,11 @@ export function TokenPicker({
   activeTokenAddress,
   onChange,
   label,
-  joined,
 }: {
   tokens: TokenChoice[];
   activeTokenAddress: Address;
   onChange: (tokenAddress: Address) => void;
   label?: string;
-  joined?: boolean;
 }): ReactElement {
   // A single element doesn't need a dropdown
   if (tokens.length === 1) {
@@ -33,11 +36,11 @@ export function TokenPicker({
             <span className="daisy-label-text">{label}</span>
           </label>
         ) : undefined}
-        <div className="daisy-join-item flex h-12 w-32 shrink-0 items-center gap-1.5 border border-neutral-content/30 bg-base-100 px-4">
+        <div className="daisy-btn no-animation daisy-btn-md flex h-9 min-h-9 cursor-default items-center rounded-box bg-neutral">
           <img
             src={tokens[0].tokenConfig.iconUrl}
             className="h-5 rounded-full"
-          />{" "}
+          />
           <span className="text-sm font-semibold">
             {tokens[0].tokenConfig.symbol}
           </span>
@@ -57,14 +60,11 @@ export function TokenPicker({
         </label>
       ) : undefined}
       <div
-        className={
-          "daisy-dropdown daisy-dropdown-bottom daisy-join-item shrink-0"
-        }
+        className={"daisy-dropdown daisy-dropdown-end daisy-dropdown-bottom"}
       >
         <button
           className={classNames(
-            "daisy-btn flex h-12 items-center border border-neutral-content/30 bg-base-100 px-4 hover:border-neutral-content/30",
-            { "border-r-none rounded-r-none": joined },
+            "daisy-btn daisy-btn-md flex h-9 min-h-9 items-center rounded-box bg-neutral pl-3 pr-1",
           )}
           onClick={(e) => {
             e.preventDefault();
@@ -75,11 +75,11 @@ export function TokenPicker({
             className="h-5 rounded-full"
           />{" "}
           {activeToken?.tokenConfig?.symbol}
-          <ChevronDownIcon className="ml-2 h-3" />
+          <ChevronDownIcon className="h-6 text-neutral-content" />
         </button>
         <ul
           className={classNames(
-            "daisy-menu daisy-dropdown-content z-[1] w-64 justify-evenly gap-0.5 rounded-lg bg-base-100 p-2 shadow",
+            "daisy-menu daisy-dropdown-content z-[1] w-72 justify-evenly gap-0.5 rounded-lg bg-base-100 p-2 shadow-md",
           )}
         >
           {[
@@ -105,7 +105,6 @@ export function TokenPicker({
                         decimals: tokenConfig?.decimals,
                         places: tokenConfig?.places,
                       })}
-                      {` `}
                     </span>
                   </label>
                 </button>
@@ -113,6 +112,115 @@ export function TokenPicker({
             )),
           ]}
         </ul>
+      </div>
+    </div>
+  );
+}
+
+export function ZapsTokenPicker({
+  tokens,
+  activeTokenAddress,
+  onChange,
+  label,
+}: {
+  tokens: TokenChoice[];
+  activeTokenAddress: Address;
+  onChange: (tokenAddress: Address) => void;
+  label?: string;
+}): ReactElement {
+  const [searchString, setSearchString] = useState("");
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  const fuse = useMemo(() => {
+    return new Fuse(tokens, {
+      keys: ["tokenConfig.symbol", "tokenConfig.name"],
+      threshold: 0.4,
+    });
+  }, [tokens]);
+
+  const filteredTokens = searchString
+    ? fuse.search(searchString).map(({ item }) => item)
+    : tokens;
+
+  const activeToken = tokens.find(
+    ({ tokenConfig }) => tokenConfig?.address === activeTokenAddress,
+  );
+
+  return (
+    <div>
+      {label ? (
+        <label className="daisy-label">
+          <span className="daisy-label-text">{label}</span>
+        </label>
+      ) : undefined}
+      <div className="daisy-dropdown daisy-dropdown-end daisy-dropdown-bottom">
+        <button
+          className={classNames(
+            "daisy-btn daisy-btn-md flex h-9 min-h-9 items-center rounded-box bg-neutral pl-3 pr-1",
+          )}
+          type="button"
+        >
+          <img
+            src={activeToken?.tokenConfig?.iconUrl}
+            className="h-5 rounded-full"
+          />{" "}
+          {activeToken?.tokenConfig?.symbol}
+          <ChevronDownIcon className="h-6 text-neutral-content" />
+        </button>
+        <div
+          tabIndex={0}
+          className="daisy-menu daisy-dropdown-content z-[1] mt-1 w-64 gap-2 rounded-lg bg-neutral p-2 shadow"
+        >
+          <div className="mb-2 flex gap-2">
+            <div className="daisy-input daisy-input-sm flex w-full items-center gap-2 rounded-md bg-base-200">
+              <input
+                ref={inputRef}
+                className="w-full shrink bg-transparent outline-none"
+                placeholder="Search"
+                value={searchString}
+                onChange={({ target }) => setSearchString(target.value)}
+              />
+              {searchString.length ? (
+                <button
+                  type="button"
+                  title="Reset search"
+                  onClick={() => {
+                    inputRef.current?.focus();
+                    setSearchString("");
+                  }}
+                  className="text-neutral-content transition-all hover:text-current"
+                >
+                  <XMarkIcon className="size-5 fill-current" />
+                </button>
+              ) : (
+                <MagnifyingGlassIcon className="size-5 fill-neutral-content" />
+              )}
+            </div>
+          </div>
+          <ul className="max-h-[40vh] overflow-y-auto">
+            {searchString.length && !filteredTokens.length ? (
+              <li className="flex h-8 items-center px-4">No matches</li>
+            ) : (
+              filteredTokens.map(({ tokenConfig }) => (
+                <li key={tokenConfig?.address}>
+                  <button
+                    type="button"
+                    onClick={() => {
+                      (document.activeElement as HTMLElement)?.blur();
+                      onChange(tokenConfig?.address);
+                    }}
+                  >
+                    <img
+                      src={tokenConfig?.iconUrl}
+                      className="h-5 rounded-full"
+                    />{" "}
+                    {tokenConfig?.symbol}
+                  </button>
+                </li>
+              ))
+            )}
+          </ul>
+        </div>
       </div>
     </div>
   );
